+++ week-08/todoapp/src/main/java/com/greenfox/todoapp/repository/TodoRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 
 
 @Component
-public interface TodoRepository extends CrudRepository<Todo, Integer> {
+public interface TodoRepository extends CrudRepository<Todo, Long> {
 
   List<Todo> findByIsDoneTrue();
